+++ src/app/chatbot-design-studio/cds-chatbot-details/detail/detail.component.ts
@@ -16,6 +16,7 @@ import { UserModel } from 'src/chat21-core/models/user';
 import { UploadModel } from 'src/chat21-core/models/upload';
 import { ImageRepoService } from 'src/chat21-core/providers/abstract/image-repo.service';
 import { checkAcceptedFile, filterImageMimeTypesAndExtensions } from '../../utils';
+import { generateSlug } from 'src/app/utils/util';
 const swal = require('sweetalert');
 
 @Component({
@@ -67,9 +68,10 @@ export class CDSDetailBotDetailComponent extends BotsBaseComponent implements On
   dept_id: string;
   selected_dept_id: string;
   selected_dept_name: string;
-
   user: UserModel
   imageURL: string;
+
+  chatbotSlugChanged: boolean;
 
   private logger: LoggerService = LoggerInstance.getInstance()
   constructor(
@@ -93,7 +95,16 @@ export class CDSDetailBotDetailComponent extends BotsBaseComponent implements On
   ngOnChanges() {
     this.logger.log('[CDS-CHATBOT-DTLS] (OnChanges) selectedChatbot ', this.selectedChatbot)
     this.destructureSelectedChatbot(this.selectedChatbot)
+    this.patchAgensAvailable();
   }
+
+
+  private patchAgensAvailable(){
+    if(this.selectedChatbot && this.selectedChatbot.agents_available != false){
+      this.selectedChatbot.agents_available = true;
+    }
+  }
+
 
 
   getDeptsByProjectId() {
@@ -512,16 +523,27 @@ export class CDSDetailBotDetailComponent extends BotsBaseComponent implements On
     })    
   }
 
+  onInputChange(event, key: string){
+    switch(key){
+      case 'slug':
+        this.selectedChatbot.slug = generateSlug((event.target as HTMLInputElement).value);
+        this.chatbotSlugChanged= true;
+        break;
+    }
+  }
+
+
   editBot() {
     // RESOLVE THE BUG 'edit button remains focused after clicking'
     this.elementRef.nativeElement.blur();
-
     // this.logger.log('[CDS-CHATBOT-DTLS] FAQ KB NAME TO UPDATE ', this.faqKb_name);
     this.faqKbService.updateFaqKb(this.selectedChatbot).subscribe({next:(faqKb) => {
       this.logger.log('[CDS-CHATBOT-DTLS] EDIT BOT - FAQ KB UPDATED ', faqKb);
       if (faqKb) {
         this.selectedChatbot.name = faqKb['name']
         this.selectedChatbot.description = faqKb['description']
+
+        this.chatbotSlugChanged = false;
       }
     }, error: (error) => {
       this.logger.error('[CDS-CHATBOT-DTLS] EDIT BOT -  ERROR ', error);
@@ -529,12 +551,33 @@ export class CDSDetailBotDetailComponent extends BotsBaseComponent implements On
       if(error && error.error.error_code === 12001){
         // =========== NOTIFY SLUG ALREADY EXISTS ERROR ===========
         this.notify.showWidgetStyleUpdateNotification(this.translationsMap.get('CDSSetting.SlugAlreadyExists'), 4, 'report_problem');
+        this.chatbotSlugChanged = false;
         return
       }
-
       // =========== NOTIFY ERROR ===========
       this.notify.showWidgetStyleUpdateNotification(this.translationsMap.get('CDSSetting.UpdateBotError'), 4, 'report_problem');
+    }, complete: () => {
+      this.logger.log('[CDS-CHATBOT-DTLS] EDIT BOT - * COMPLETE *');
+      // =========== NOTIFY SUCCESS===========
+      this.notify.showWidgetStyleUpdateNotification(this.translationsMap.get('CDSSetting.UpdateBotSuccess'), 2, 'done');
+    }});
+  }
 
+
+  editBotAgentsAvailable() {
+    // RESOLVE THE BUG 'edit button remains focused after clicking'
+    this.elementRef.nativeElement.blur();
+    // this.logger.log('[CDS-CHATBOT-DTLS] FAQ KB NAME TO UPDATE ', this.faqKb_name);
+    this.faqKbService.updateFaqKbAgentsAvailable(this.selectedChatbot._id, this.selectedChatbot.agents_available).subscribe({next:(faqKb) => {
+      this.logger.log('[CDS-CHATBOT-DTLS] EDIT BOT - FAQ KB UPDATED ', faqKb);
+      if (faqKb) {
+        this.selectedChatbot.name = faqKb['name']
+        this.selectedChatbot.description = faqKb['description']
+      }
+    }, error: (error) => {
+      this.logger.error('[CDS-CHATBOT-DTLS] EDIT BOT -  ERROR ', error);
+      // =========== NOTIFY ERROR ===========
+      this.notify.showWidgetStyleUpdateNotification(this.translationsMap.get('CDSSetting.UpdateBotError'), 4, 'report_problem');
     }, complete: () => {
       this.logger.log('[CDS-CHATBOT-DTLS] EDIT BOT - * COMPLETE *');
       // =========== NOTIFY SUCCESS===========
@@ -542,6 +585,7 @@ export class CDSDetailBotDetailComponent extends BotsBaseComponent implements On
       this.selectedChatbot.name
     }});
   }
+
 
   goToRoutingAndDepts() {
     let redirecturl = this.appConfigService.getConfig().dashboardBaseUrl + '#/project/'+ this.project._id + '/departments'
@@ -552,5 +596,15 @@ export class CDSDetailBotDetailComponent extends BotsBaseComponent implements On
     let redirecturl = this.appConfigService.getConfig().dashboardBaseUrl + '#/project/'+ this.project._id + '/department/edit/'+ deptid
     window.open(redirecturl, '_blank')
   }
+
+
+  onAgentsAvailableChange(event: Event): void {
+    const checkbox = event.target as HTMLInputElement;
+    const isAgentsAvailable = checkbox.checked;
+    this.logger.log('[CDS-CHATBOT-DTLS] isAgentsAvailable: ',isAgentsAvailable);
+    this.selectedChatbot.agents_available = isAgentsAvailable;
+    this.editBotAgentsAvailable();
+  }
+
 
 }