+++ src/app/chatbot-design-studio/cds-dashboard/cds-canvas/cds-intent/panel-intent-header/panel-intent-header.component.ts
@@ -31,13 +31,14 @@ export class PanelIntentHeaderComponent implements OnInit, OnChanges {
     this.intentService.getIntents().subscribe(intents => {
       if(intents){
         this.listOfIntents = intents;
+        this.intentName = this.intent.intent_display_name;
       }
     })
   }
 
   // SYSTEM FUNCTIONS //
   ngOnInit(): void {
-    this. initialize();
+    this.initialize();
   }
 
   ngOnChanges() {